4.2 Versão

+++ VisuDados/models.py
@@ -5,7 +5,7 @@ from django.db import models
 class Data(models.Model):
     dado = models.FloatField()
     dado2 = models.FloatField()
-    label = models.IntegerField()
+    label = models.FloatField()
 
     def __str__(self):
         return f"{self.label}"
